+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+## [0.20.0] - 2018-03-18
+
 ### Added
 
 - metrics/prometheus: Prometheus instance
@@ -325,7 +328,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Bootstrap options
 
 
-[Unreleased]: https://github.com/goph/fxt/compare/v0.19.1...HEAD
+[Unreleased]: https://github.com/goph/fxt/compare/v0.20.0...HEAD
+[0.20.0]: https://github.com/goph/fxt/compare/v0.19.1...v0.20.0
 [0.19.1]: https://github.com/goph/fxt/compare/v0.19.0...v0.19.1
 [0.19.0]: https://github.com/goph/fxt/compare/v0.18.0...v0.19.0
 [0.18.0]: https://github.com/goph/fxt/compare/v0.17.0...v0.18.0